// src/core/mips/single-cycle/diagram/types.ts
export type DatapathPath =
    | 'PC_TO_IM'
    | 'PC_TO_ADD4'
    | 'CONST4_TO_ADD4'
    | 'ADD4_TO_BRANCH_ADDER0'
    | 'ADD4_TO_PCSRC_MUX0'
    | 'PCSRC_MUX_TO_PC'
    | 'IM_TO_IR'
    | 'IR_RS_TO_RF_RR1'
    | 'IR_RT_TO_RF_RR2'
    | 'IR_RT_TO_REGDST_MUX0'
    | 'IR_RD_TO_REGDST_MUX1'
    | 'REGDST_MUX_TO_RF_WR'
    | 'IR_IMM_TO_SIGN_EXTEND'
    | 'RF_RD1_TO_ALU1'
    | 'RF_RD2_TO_ALUSRC_MUX0'
    | 'SIGN_EXTEND_TO_ALUSRC_MUX1'
    | 'ALUSRC_MUX_TO_ALU2'
    | 'SIGN_EXTEND_TO_LEFT_SHIFT_2'
    | 'LEFT_SHIFT_2_TO_BRANCH_ADDER1'
    | 'BRANCH_ADDER_TO_PCSRC_MUX1'
    | 'ALU_TO_DM_ADDR'
    | 'RF_RD2_TO_DM_WD'
    | 'ALU_ZERO_TO_BEQ_AND_GATE'
    | 'BRANCH_TO_BEQ_AND_GATE'
    | 'ALU_ZERO_TO_NOT_GATE'
    | 'NOT_GATE_TO_BNE_AND_GATE'
    | 'BRANCHNE_TO_BNE_AND_GATE'
    | 'BEQ_AND_GATE_TO_OR_GATE'
    | 'BNE_AND_GATE_TO_OR_GATE'
    | 'OR_GATE_TO_PCSRC_MUX'
    | 'ALU_TO_MEMTOREG_MUX0'
    | 'DM_RD_TO_MEMTOREG_MUX1'
    | 'MEMTOREG_MUX_TO_RF_WD'
    | 'IR_OPCODE_TO_CONTROL'
    | 'CONTROL_TO_REGDST_MUX'
    | 'CONTROL_TO_ALUSRC_MUX'
    | 'CONTROL_TO_MEMTOREG_MUX'
    | 'CONTROL_TO_REGWRITE'
    | 'CONTROL_TO_MEMREAD'
    | 'CONTROL_TO_MEMWRITE'
    | 'IR_FUNCT_TO_ALU_CONTROL'
    | 'ALUOP_TO_ALU_CONTROL'
    | 'ALU_CONTROL_TO_ALU';

export type DatapathSegment =
    | 'PC_TO_PC_JUNCTION'
    | 'PC_JUNCTION'
    | 'PC_JUNCTION_TO_ADD4'
    | 'PC_JUNCTION_TO_IM'
    | 'ADD4_TO_ADD4_JUNCTION'
    | 'ADD4_JUNCTION'
    | 'ADD4_JUNCTION_TO_BRANCH_ADDER0'
    | 'ADD4_JUNCTION_TO_PCSRC_MUX0'
    | 'CONST4_TO_ADD4'
    | 'PCSRC_MUX_TO_PC'
    | 'IM_TO_IR'
    | 'IR_RS_TO_RF_RR1'
    | 'IR_RT_TO_RT_JUNCTION'
    | 'RT_JUNCTION'
    | 'RT_JUNCTION_TO_RF_RR2'
    | 'RT_JUNCTION_TO_REGDST_MUX0'
    | 'IR_RD_TO_REGDST_MUX1'
    | 'REGDST_MUX_TO_RF_WR'
    | 'IR_IMM_TO_SIGN_EXTEND'
    | 'RF_RD1_TO_ALU1'
    | 'RF_RD2_TO_RD2_JUNCTION'
    | 'RD2_JUNCTION'
    | 'RD2_JUNCTION_TO_ALUSRC_MUX0'
    | 'SIGN_EXTEND_TO_SE_JUNCTION'
    | 'SE_JUNCTION'
    | 'SE_JUNCTION_TO_SE_JUMP1'
    | 'SE_JUMP1'
    | 'SE_JUMP1_TO_SE_JUMP2'
    | 'SE_JUMP2'
    | 'SE_JUMP2_TO_LEFT_SHIFT_2'
    | 'SE_JUNCTION_TO_ALUSRC_MUX1'
    | 'ALUSRC_MUX_TO_ALU2'
    | 'LEFT_SHIFT_2_TO_BRANCH_ADDER1'
    | 'BRANCH_ADDER_TO_PCSRC_MUX1'
    | 'ALU_TO_ALU_JUNCTION'
    | 'ALU_JUNCTION'
    | 'ALU_JUNCTION_TO_DM_ADDR'
    | 'RD2_JUNCTION_TO_RD2_JUMP'
    | 'RD2_JUMP'
    | 'RD2_JUMP_TO_DM_WD'
    | 'ALU_JUNCTION_ALU_JUMP'
    | 'ALU_JUMP'
    | 'ALU_JUMP_TO_MEMTOREG_MUX0'
    | 'DM_RD_TO_MEMTOREG_MUX1'
    | 'MEMTOREG_MUX_TO_MEMTOREG_MUX_JUMP'
    | 'MEMTOREG_MUX_JUMP'
    | 'MEMTOREG_MUX_JUMP_TO_RF_WD';

export type DatapathValueId =
    | 'PC'
    | 'ADD4'
    | 'BRANCH_ADDER'
    | 'SIGN_EXTEND'
    | 'LEFT_SHIFT_2'
    | 'IM_ADDRESS'
    | 'IM_INSTRUCTION'
    | 'IR_OPCODE'
    | 'IR_RS'
    | 'IR_RT'
    | 'IR_RD'
    | 'IR_SHAMT'
    | 'IR_FUNCT'
    | 'IR_IMMEDIATE'
    | 'RF_RR1'
    | 'RF_RR2'
    | 'RF_WR'
    | 'RF_RD1'
    | 'RF_RD2'
    | 'RF_WD'
    | 'ALU_OP1'
    | 'ALU_OP2'
    | 'ALU_RESULT'
    | 'ALU_ZERO'
    | 'DM_ADDRESS'
    | 'DM_WRITE_DATA'
    | 'DM_READ_DATA';